✨ feat(Added and authentication middleware which uses JWT secrets):
This requires you to configure the environment variables with the correct values which you can obtain from the supabase console. The authCheck function that I have added is used as a middleware for those api routes that needs an authenticated user, those routes that do not then they won't use the authCheck function.