✨ feat(Created new module): `user` module
I have created the user module, where controllers, routes and so on will be placed. At the moment I have created a few models that appartain to the the user module, namely `User`, `UserProfile`, and `Feedback`. I have also fixed an issue with the authCheck, which was that I misconfigured the config and environment variables.