✨ feat(User can now be created following successful authentication): User Creation
User and its following associated tables can now be created when a successful authentication has been made. Moreover, to make the  possible more folder like validation, controller and so on were created to aid in this process as well as making the code modularised.Path alias are now used for easier access to files.